refactor: Move Application Coroutine Scope provider from app to core:common module
