+++ app/src/main/java/com/octopus/edu/trackmate/di/AppModule.kt
@@ -2,30 +2,16 @@ package com.octopus.edu.trackmate.di
 
 import android.content.Context
 import androidx.work.WorkManager
-import com.octopus.edu.core.common.DispatcherProvider
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.SupervisorJob
-import javax.inject.Qualifier
 import javax.inject.Singleton
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class ApplicationScope
 
 @Module
 @InstallIn(SingletonComponent::class)
 object AppModule {
-    @Singleton
-    @ApplicationScope
-    @Provides
-    fun providesApplicationScope(dispatcherProvider: DispatcherProvider): CoroutineScope =
-        CoroutineScope(SupervisorJob() + dispatcherProvider.default)
-
     @Singleton
     @Provides
     fun providesWorkManager(
