+++ app/src/main/java/com/octopus/edu/trackmate/sync/EntrySyncManager.kt
@@ -3,11 +3,11 @@ package com.octopus.edu.trackmate.sync
 import com.octopus.edu.core.common.DispatcherProvider
 import com.octopus.edu.core.common.Logger
 import com.octopus.edu.core.common.RetryPolicy
+import com.octopus.edu.core.common.di.ApplicationScope
 import com.octopus.edu.core.data.entry.di.SyncErrorClassifierQualifier
 import com.octopus.edu.core.domain.repository.EntryRepository
 import com.octopus.edu.core.domain.scheduler.EntrySyncScheduler
 import com.octopus.edu.core.domain.utils.ErrorClassifier
-import com.octopus.edu.trackmate.di.ApplicationScope
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.catch
